Removed BazaarComponents folder
