+++ plugins/bazaar/dev/index.tsx
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import React from 'react';
 import { createDevApp } from '@backstage/dev-utils';
 import { bazaarPlugin, BazaarPage } from '../src/plugin';
