+++ plugins/bazaar/src/components/AddProjectDialog/AddProjectDialog.tsx
@@ -33,8 +33,8 @@ import { Entity } from '@backstage/catalog-model';
 import { useApi, githubAuthApiRef } from '@backstage/core-plugin-api';
 import { useAsync } from 'react-use';
 import { InputMultiSelector } from '../InputMultiSelector';
-import { createPullRequest, getBranches } from '../../../util/githubUtils';
-import { editBazaarProperties } from '../../../util/editBazaarProperties';
+import { createPullRequest, getBranches } from '../../util/githubUtils';
+import { editBazaarProperties } from '../../util/editBazaarProperties';
 import { InputField } from '../InputField';
 import { InputSelector } from '../InputSelector';
 import { ProjectSelector } from '../ProjectSelector';
