+++ plugins/bazaar/src/components/AddProjectDialog/index.ts
