+++ plugins/bazaar/src/components/BazaarInfoCard/BazaarInfoCard.test.tsx
