+++ plugins/bazaar/src/components/DeleteProjectDialog/index.ts
