+++ plugins/bazaar/src/components/HomePage/HomePage.test.tsx
