+++ plugins/bazaar/src/components/InputSelector/InputSelector.tsx
