+++ plugins/bazaar/src/components/InputSelector/index.ts
