+++ plugins/bazaar/src/components/ProjectCard/ProjectCard.tsx
