+++ plugins/bazaar/src/components/ProjectCard/index.ts
