+++ plugins/bazaar/src/components/ProjectCatalog/ProjectCatalog.tsx
