+++ plugins/bazaar/src/components/ProjectCatalog/index.ts
