+++ plugins/bazaar/src/components/ProjectPreview/ProjectPreview.tsx
