+++ plugins/bazaar/src/components/ProjectPreview/index.ts
