+++ plugins/bazaar/src/components/ProjectSelector/ProjectSelector.tsx
