+++ plugins/bazaar/src/components/ProjectSelector/index.ts
